% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expobiolink.R
\name{FuncExit}
\alias{FuncExit}
\title{End the module analysis}
\usage{
FuncExit(PID)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by any initial functions.}
}
\value{
Exit status
}
\description{
End the module analysis
}
\examples{
res = InitBioLink()
   res = LoadBioLink(PID = res$PID,UseExample = "example#1")
   res2 = ConvToExpoID(PID = res$PID)
   FuncExit(PID = res$PID)
}
\author{
Bin Wang (corresponding author)
}
